tidy up 2nd draft, but print left align space has to be fixed

+++ src/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: hokutosuzuki <hosuzuki@student.42toky      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/30 17:23:17 by hokutosuz         #+#    #+#             */
-/*   Updated: 2022/02/02 21:10:32 by hokutosuz        ###   ########.fr       */
+/*   Updated: 2022/02/03 17:44:52 by hokutosuz        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int	ft_printf(const char *fmt, ...)
 {
 	t_stock	*lst;
-	size_t	total_length;
+	int		total_length;
 
 	if (!fmt)
 		return (0);
